requirements.txt
google-cloud
google-cloud-translate
google-api-python-client
PyHamcrest
python-dotenv
google
